pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "fcut"
version = "0.0.7"
description = "FCut Video Processing Application"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ffmpeg-toolkit>=0.2.0",
    "flask>=3.1.0",
    "pydantic>=1.10.21",
    "pywebview>=5.4",
]

[dependency-groups]
dev = [
    "nuitka>=2.6.9",
]
